Shop-18: Better active link, feedback about product with error/success sending window
