+++ resources/views/category.blade.php
@@ -1,4 +1,4 @@
-@extends('welcome')
+@extends('layouts.base')
 @section('content')
 
 <body id="category">
