+++ resources/views/main.blade.php
@@ -1,4 +1,4 @@
-@extends('welcome')
+@extends('layouts.base')
 @section('content')
 
     <!-- start banner Area -->
