+++ resources/views/tracking.blade.php
@@ -1,4 +1,4 @@
-@extends('welcome')
+@extends('layouts.base')
 @section('content')
 
     <!-- Start Banner Area -->
